refactor: split epbf.py into smaller, focused modules
Address code review comment to improve maintainability by splitting
the large epbf.py file into smaller, single-responsibility modules.

New Structure:
- **source_query_proxy/ebpf/runtime.py** (230 lines)
  - Network interface helpers (get_addr_interface, get_default_interface)
  - Server configuration collection (collect_server_mappings)
  - BPF program generation (generate_bpf_program)
  - Map population (populate_maps, ip_to_int)
  - BCC lazy import (import_bcc)

- **source_query_proxy/ebpf/tc.py** (200 lines)
  - TC attachment logic (attach_tc_bpf)
  - Rollback on failure (rollback_tc_bpf)
  - TC cleanup (cleanup_tc)
  - All traffic control operations in one place

- **source_query_proxy/ebpf/redirector.py** (210 lines)
  - EBPFRedirector class (async lifecycle manager)
  - start(), stop(), restart() methods
  - Async context manager support
  - Clean separation of concerns

- **source_query_proxy/epbf.py** (105 lines, was 630 lines)
  - Main entry point (run_ebpf_redirection)
  - Backward compatibility (get_ebpf_program_run_args)
  - Re-exports for compatibility
  - Clear documentation of new structure

Benefits:
- 84% reduction in main file size (630 → 105 lines)
- Clear separation of concerns (network, TC, lifecycle)
- Easier to test individual components
- Better discoverability
- Simpler import paths for specific functionality
- Follows single-responsibility principle

Usage (unchanged for backward compatibility):
    # Old way still works
    from source_query_proxy.epbf import run_ebpf_redirection
    await run_ebpf_redirection()

    # New way (recommended)
    from source_query_proxy.ebpf.redirector import EBPFRedirector
    async with EBPFRedirector() as redirector:
        await asyncio.Event().wait()

All tests and existing code continue to work without modification.
